Removing core migration history bookmarks passwords and version updates
