--- Sources/Brave/Frontend/Sync/WebFilter/Bookmarks/Bookmarks.html
@@ -1,101 +1,0 @@
-<!--
-Copyright (c) 2020 The Brave Authors. All rights reserved.
-This Source Code Form is subject to the terms of the Mozilla Public
-License, v. 2.0. If a copy of the MPL was not distributed with this file,
-You can obtain one at https://mozilla.org/MPL/2.0/.
--->
-
-<head>
-    <meta http-equiv="Content-Security-Policy" content="default-src 'self'">
-    <meta name='viewport' content='initial-scale=1, viewport-fit=cover'>
-    <meta name="referrer" content="no-referrer">
-    <title>Bookmarks</title>
-    <style>
-        .post {
-            padding-top: max(25px, env(safe-area-inset-top));
-            padding-bottom: max(25px, env(safe-area-inset-bottom));
-            padding-left: max(25px, env(safe-area-inset-left));
-            padding-right: max(25px, env(safe-area-inset-right));
-        }
-
-        .background {
-            background-color: white;
-        }
-
-        .icon {
-            width: 64px;
-            height: 64px;
-        }
-
-        .title {
-            color: rgb(94, 97, 117);
-            font-size: 20px;
-            font-family: SFProDisplay-Medium, -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Open Sans', 'Helvetica Neue', sans-serif;
-            font-weight: 500;
-            letter-spacing: 0.01px;
-            line-height: 28px;
-        }
-
-        .description {
-            color: rgb(82, 82, 82);
-            font-size: 14px;
-            font-family: SFProText-Regular, -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Open Sans', 'Helvetica Neue', sans-serif;
-            font-weight: normal;
-            letter-spacing: 0px;
-            line-height: 19px;
-        }
-
-        .container {
-            display: flex;
-            flex-wrap: wrap;
-            flex-direction: column;
-            align-content: flex-start;
-        }
-
-        /** Center the content for iPads **/
-        @media (min-width: 600px) and (min-height: 600px) {
-            .content {
-                margin: 0;
-                position: absolute;
-                top: 50%;
-                left: 50%;
-                max-width: 650px;
-                -ms-transform: translate(-50%, -50%);
-                transform: translate(-50%, -50%);
-            }
-        }
-    
-        h1 {
-            visibility: hidden;
-        }
-    </style>
-</head>
-
-<body class="background content">
-    <div class="container post">
-        <svg class="icon" xmlns="http://www.w3.org/2000/svg">
-            <defs>
-                <linearGradient x1=".643%" y1="49.357%" x2="99.357%" y2="49.357%" id="a">
-                    <stop stop-color="#392DD1" offset="0%" />
-                    <stop stop-color="#A91B78" offset="100%" />
-                </linearGradient>
-            </defs>
-            <g fill="none" fill-rule="evenodd">
-                <path d="M0 0h64v64H0z" />
-                <path
-                    d="M55.417 7A4.59 4.59 0 0160 11.583v41.25a4.59 4.59 0 01-4.583 4.584H9.583A4.59 4.59 0 015 52.833v-41.25A4.59 4.59 0 019.583 7h45.834zm-.001 13.75H9.583v32.083h45.834l-.001-32.083zM32.5 36.768c3.571 0 7.066 1.464 10.45 4.298a2.279 2.279 0 01.29 3.194 2.242 2.242 0 01-3.172.293c-2.623-2.198-5.133-3.25-7.568-3.25-2.435 0-4.945 1.052-7.568 3.25a2.242 2.242 0 01-3.172-.293 2.279 2.279 0 01.29-3.194c3.384-2.834 6.879-4.298 10.45-4.298zm-6.756-9.071a2.26 2.26 0 012.252 2.267 2.26 2.26 0 01-2.252 2.268 2.26 2.26 0 01-2.253-2.268 2.26 2.26 0 012.253-2.267zm13.512 0a2.26 2.26 0 012.253 2.267 2.26 2.26 0 01-2.253 2.268 2.26 2.26 0 01-2.252-2.268 2.26 2.26 0 012.252-2.267zm16.16-16.114H9.584v4.583h45.833v-4.583z"
-                    fill="url(#a)" />
-            </g>
-        </svg>
-
-        <p class="title">
-            %bookmarks_tab_title%
-        </p>
-
-        <p class="description">
-            %bookmarks_page_description%
-            <br /><br />
-            %bookmarks_file_url%
-        </p>
-    </div>
-</body>
